stm32mp1: add trusted boot with TF-A
Add support of trusted boot, using TF-A as first stage bootloader,
The boot sequence is
  BootRom >=> TF-A.stm32 (clock & DDR) >=> U-Boot.stm32

The TF-A monitor provides secure monitor with support of SMC
- proprietary to manage secure devices (BSEC for example)
- PSCI for power

The same device tree is used for STMicroelectronics boards with
basic boot and with trusted boot.

Series-version: 4

Series-changes: 4
    - Rebase on master
    - add CONFIG_CMD_CLK in trusted defconfig
      (allignment with basic defconfig)

Series-changes: 3
   - Remove patch 4/4 added by error in v2 and sent by other serie
     http://patchwork.ozlabs.org/patch/1033475/

Series-changes: 2
    - Rebase and solve conflict on arch/arm/mach-stm32mp/config.mk

Cover-letter:
stm32mp1: add trusted boot with TF-A

This patch serie add a new configuration
stm32mp15_trusted_defconfig to handle STM32MP157 boards
with TF-A as secure first stage bootloader.

TF-A code is available in GitHub
https://github.com/ARM-software/arm-trusted-firmware

With platform stm32mp1 platform
https://github.com/ARM-software/arm-trusted-firmware/tree/master/plat/st/stm32mp1

For details, see documentation
https://github.com/ARM-software/arm-trusted-firmware/blob/master/docs/plat/stm32mp1.rst

PS: upstream of trusted boot for TF-A is still in progress.

END

Change-Id: Iba60f508d74d47fa2ff2be9d616ed28a5864fdc4
Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>