Convert CONFIG_ENV_SPI_* to Kconfig
This converts the following to Kconfig:
  CONFIG_ENV_SPI_BUS
  CONFIG_ENV_SPI_CS
  CONFIG_ENV_SPI_MAX_HZ
  CONFIG_ENV_SPI_MODE

Most of time these value are not needed, CONFIG_SF_DEFAULT
with same value is used, so I introduced CONFIG_USE_ENV_SPI_*
to force the associated value for the environment.

Cover-letter:
Migrate SPI defines to Kconfig

This serie is a proposal to convert all SPI configuration settings
in Kconfig :
+ parameters for SF command
  - CONFIG_SF_DEFAULT_BUS
  - CONFIG_SF_DEFAULT_CS
  - CONFIG_SF_DEFAULT_MODE
  - CONFIG_SF_DEFAULT_SPEED
+ parameters for SSPI command
  - CONFIG_DEFAULT_SPI_BUS
  - CONFIG_DEFAULT_SPI_MODE
+ parameters for ENV in SPI
  - CONFIG_ENV_SPI_BUS
  - CONFIG_ENV_SPI_CS
  - CONFIG_ENV_SPI_MAX_HZ
  - CONFIG_ENV_SPI_MODE

I make this work after the remarks of Jagan Teki.
on http://patchwork.ozlabs.org/patch/1031768/
[U-Boot,v3] dm: spi: Read default speed and mode values from DT

So it is a preliminary for v4 on this patch.

This commit includes a full migration using moveconfig.py to ensure
that each commit compiles.

To allow compilation, I also move CMD in Kconfig for some board.

I introduce CONFIG_USE_ENV_SPI* to handle absent value
(CONFIG_SF_DEFAULT_* is used) and use some post-processing to solve
the remaining issue and remove the unnecessary configuration in defconfig
(when default value can be used).
I make manual check at the end.

Compilation is OK on Travis:
https://travis-ci.org/patrickdelaunay/u-boot/builds/498044826

This v1 can be see as a starting point based on v2019.04-rc2.
Any feedback or test is welcome.

END

Change-Id: If5c814274e60cd42f7a40ab8c4a814257a735759
Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>