stm32mp1: add stusb1600 support for DK1 and DK2 board
The DK1 and DK2 boards use the USB Type-C controller STUSB1600.
This patch updates:
- the device tree to add the I2C node in the DT
- the board stm32mp1 to probe this I2C device and use this controller
  to check cable detection.
- the DWC2 driver to support a new dt property
  "u-boot,force-b-session-valid" which forces B session and
  device mode; it is a workaround because the VBUS sensing and
  ID detection isn't available with stusb1600.

Cover-letter:
usb: convert dwc2 gadget to driver model, used in stm32mp1

This patch-set created after Marek remarks on patch
  board: stm32mp1: Add tx_fifo_sz_array support
  http://patchwork.ozlabs.org/patch/1056452/

This serie convert the DWC2 to driver model and use it for the
stm32mp1 boards.

USB device and cable connection tested on EV1 and DK2.

This serie superseded the Patrice patches:
http://patchwork.ozlabs.org/project/uboot/list/?series=97121

Warning: the serie need to be apply after
http://patchwork.ozlabs.org/project/uboot/list/?series=91422
http://patchwork.ozlabs.org/project/uboot/list/?series=91497

END

Change-Id: I305537af6a038a91e00607a3131995dea7baac4a
Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>