stm32mp1: add support for virtual partition read
Add read for OTP and PMIC NVM with alternates
on virtual DFU device.

Series-version: 3

Series-version: 3
- rebase on v2019.10

Series-changes: 2
- Update after Lukasz Majewski comments

Cover-letter:
dfu: update dfu stack and add MTD backend

This serie based on v2019.10 propose updates on the DFU stack:
- add capability to have several DFU backend running in parallel
- add MTD backend for update of NAND, NOR or SPI-NAND
- add VIRTUAL backend for board/command specific behavior
- add some weak callback

It is the same content that the previous v2:
+ solve compilation for dependancy with CONFIG_CMD_MTDPARTS
+ solve bad block management for NAND in the erase loops
  (copied from MTD command), the previous was functional only on
  SPI-NAND. But I detect issue for NAND; when bad block is detected
  fail_addr is not updated.
  See the complete analysis on the issue (reproduced in MTD command)
  in http://patchwork.ozlabs.org/patch/1165056/
  "cmd: mtd: solve bad block support in erase command"
  The better solution is a MTD backport from Linux, but udpate the
  loop is functional for SPI-NAND and NAND.

To test the new features and as example, I update the stm32mp1 board
to use the new features; I test them with the command "dfu 0"
(for test I have dependency with patch
 http://patchwork.ozlabs.org/patch/1162076/)

The expected target of the dfu part for this serie is v2020.01.
I will merge the stm32mp1 part when they will be accepted by dfu
maintainers.

This serie prepares the DFU backend for the coming 'stm32prog' command
and for STM32CubeProgrammer on stm32mp1 platform.
This STMicroelectronics tool is based on DFU protocol and allows to
update all the boot devices and the OTPs on the ST boards
(so several DFU target in parallel, including spi nand).

Build result dfu_v3
https://travis-ci.org/patrickdelaunay/u-boot/builds/596626418

Regards

Patrick

END

Serie-cc: Boris Brezillon <boris.brezillon@bootlin.com>
Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>
Change-Id: Iaffc9f74dba9fc929fc3e4dbe40236f00a744e03