env: spl: filter the variables in default environment of SPL or TPL
Use a new option CONFIG_SPL_ENV_VARS to filter the variables included
in the default environment used in SPL (and TPL).

That allows to reduce the size of default_environment[].

If the new configuration flags CONFIG_SPL_ENV_VARS is
- "" the default environment is empty (default)
- list of variables (comma separated) to keept in SPL/TPL
  (make search ^<var>= in u-boot-initial-env)
- "*" the variables are not filtered

Series-prefix: RFC RFT

Commit-notes:
Hi,

I propose this patch to reduce the SPL/TPL size when they support the
U-Boot environment is activated.

But I need review and test (mainly for first boot when environment
in storage device is empty) to confirm my assumptions on the list
of variable used in SPL .

This patch adds a filter on the environment variable present
in default environment of U-Boot: SPL/TPL only include the requested one.

I assumed the SPL/TPL can be empty by default (I filter all the variables),
the environment is only needed when it is loaded from external storage
device.
For me on cold boot and before to save save, the content of default
environment it is not necessary.

For example, with falcon mode we need the environment to have the bootcmd
(load address , the file name to load and the the bootargs) only after
a first boot and a correct env save.

But I can update the default value for CONFIG_SPL_ENV_VARS to add
some generic variable to solve issues:
	.flags
	.callback
	cpu
	arch
	vendor
	soc

PS: the current behavior (no filter) is kept when
    CONFIG_SPL_ENV_VARS = "*"

I propose this patch after remarks on previous patch v4
env: Add CONFIG_ENV_FULL_SUPPORT

http://patchwork.ozlabs.org/patch/1171180/

It is a preliminary step for v5 of this serie.

Patrick

END

Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>