gpio: update dir_flags management
Update the flag in GPIO uclass: the desc->flags is always
combined with the requested flags and GPIO descriptor is
updated for further call.

Add a function dm_gpio_get_dir_flags to get the dynamically
the current dir_flags (configuration and value).

This patch prepare introduction of new dir flags support.

Series-changes: 3
- part of v2 08/14 = gpio: add ops for configuration with dir flags

Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>